Cleanup other.test_single_file. NFC
- Use `also_with_wasm2js` decorator
- Peel one of the loop variables and implement it as @parameterize
  instead.
